- [3dfloors branch] - Switched from a 32-bit z-buffer to a 16-bit inverse z-buffer. - Fixed: the texture shifts for obscured parts of a plane didn't account for multiple obstructions. - Added segs and things to the inverse z-buffer, still a bit glitchy and I don't think I accounted for sloped planes yet. - Attempted to reverse the order in which fake planes are rendered.
SVN r2443 (3dfloors)

+++ src/r_bsp.cpp
@@ -84,7 +84,7 @@ int fake3D;
 F3DFloor *fakeFloor;
 fixed_t fakeHeight;
 fixed_t fakeAlpha;
-DWORD* zbuffer = NULL;
+WORD* zbuffer = NULL;
 bool useZBuffer;
 
 fixed_t			rw_backcz1, rw_backcz2;
